% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tax_from_ncbi.R
\name{download_tax_from_ncbi}
\alias{download_tax_from_ncbi}
\title{Create a dataframe with taxonomic information extracted from NCBI.}
\usage{
download_tax_from_ncbi(
  taxid,
  rank_vc = c("superkingdom", "kingdom", "subphylum", "superclass", "class",
    "suborder", "order", "family", "subfamily", "genus")
)
}
\arguments{
\item{taxid}{A character vector containing Uniprot taxonomy id}

\item{rank_vc}{A character vector containing Taxonomic rank columns}
}
\value{
Dataframe
}
\description{
Create a dataframe with taxonomic information extracted from NCBI.
}
\examples{
\dontrun{
#sequential
download_xml_from_phmer(body_list =
    list(seqdb = "pdb",seq = '>Seq\nKLRVLGYH'))
}
}
