% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_xml_from_phmer.R
\name{download_xml_from_phmer}
\alias{download_xml_from_phmer}
\title{download_xml_from_phmer}
\usage{
download_xml_from_phmer(body_list)
}
\arguments{
\item{body_list}{A list with body parameters for the POST request.}
}
\value{
A XML file in as raw.
}
\description{
download_xml_from_phmer
}
\examples{
\dontrun{
download_xml_from_phmer(body_list =
    list(seqdb = "pdb",seq = '>Seq\nKLRVLGYH'))
}
}
