% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SCA_positional_coevolution_matrix_eigenvalues_data.R
\name{eigenspectrum_SCA_positional_coevolution_matrix_plot}
\alias{eigenspectrum_SCA_positional_coevolution_matrix_plot}
\title{Plot the eigenspectrum of the SCA positional coevolution matrix (Cij~)
and 10 trials of matrix randomization for comparison. This graph is used
to choose the number of significant eigenmodes}
\usage{
eigenspectrum_SCA_positional_coevolution_matrix_plot(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
Plot the eigenspectrum of the SCA positional coevolution matrix (Cij~)
and 10 trials of matrix randomization for comparison. This graph is used
to choose the number of significant eigenmodes
}
