% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_df_from_hmmer.R
\name{extract_tidy_df_from_hmmer}
\alias{extract_tidy_df_from_hmmer}
\title{extract_tidy_df_from_hmmer}
\usage{
extract_tidy_df_from_hmmer(
  xml.document,
  by_column = c(alisqacc = "acc", alisqname = "name")
)
}
\arguments{
\item{xml.document}{A xml_document downloaded from HMMER}

\item{by_column}{A character vector for joining domains hash with sequence's hits hash.}
}
\value{
DataFrame
}
\description{
extract_tidy_df_from_hmmer
}
\examples{
\dontrun{
 xml.path \%>\%
   read_xml() \%>\%
   extract_tidy_df_from_hmmer()
}
}
