% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tidy_stats_from_xml.R
\name{extract_tidy_stats_from_xml}
\alias{extract_tidy_stats_from_xml}
\title{extract_tidy_stats_from_xml}
\usage{
extract_tidy_stats_from_xml(xml.document)
}
\arguments{
\item{xml.document}{A xml_document downloaded from HMMER that contain stats}
}
\value{
A tidy data.frame containing available stats' information in xml_document.
\itemize{
\item{nhits: The number of hits found above reporting thresholds}
\item{Z: The number of sequences or models in the target database}
\item{domZ: The number of hits in the target database}
\item{nmodels: The number of models in this search}
\item{nincluded: The number of sequences or models scoring above the significance threshold}
\item{nreported: The number of sequences or models scoring above the reporting threshold}
}
}
\description{
extract_tidy_stats_from_xml
}
\examples{
\dontrun{
read_xml("2abl_A_pdb.xml") \%>\%
    extract_tidy_stats_from_xml()
}
}
