% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_similarity_matrix_data.R
\name{global_similarity_matrix_data}
\alias{global_similarity_matrix_data}
\title{It computes the data from the similarity matriz in order to visualize it.}
\usage{
global_similarity_matrix_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Tidy DataFrame
}
\description{
It computes the data from the similarity matriz in order to visualize it.
}
