% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_similarity_matrix_data.R
\name{global_similarity_matrix_heatmap}
\alias{global_similarity_matrix_heatmap}
\title{Plot a heatmap in order to visualize
global view of the sequence similarity matrix}
\usage{
global_similarity_matrix_heatmap(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
Plot a heatmap in order to visualize
global view of the sequence similarity matrix
}
