% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pySCA_module_using_reticulate.R
\name{import_pySCA_module_using_reticulate}
\alias{import_pySCA_module_using_reticulate}
\title{import_pySCA_module_using_reticulate}
\usage{
import_pySCA_module_using_reticulate(
  virtualenv = NULL,
  condaenv = NULL,
  force_installing = FALSE
)
}
\arguments{
\item{virtualenv}{Either the name of, or the path to, a Python virtual environment. NULL if you do not want to specify.}

\item{condaenv}{The name of the Conda environment to use. NULL if you do not want to specify.}

\item{force_installing}{A boolean, if true, will create an env  and install missing python modules.}
}
\value{
pySCA module
}
\description{
import_pySCA_module_using_reticulate
}
\examples{
\dontrun{
library(reticulate)
use_virtualenv("~/myenv")
use_condaenv("myenv")
pysca <- import_pySCA_module_using_reticulate
pysca$scaTools
}
}
