% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_sequence_identities_data.R
\name{pairwise_sequence_identities_data}
\alias{pairwise_sequence_identities_data}
\title{It computes the data from the simmilarity matrix in order to visualize
the pairwise sequence identities}
\usage{
pairwise_sequence_identities_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Tidy DataFrame
}
\description{
It computes the data from the simmilarity matrix in order to visualize
the pairwise sequence identities
}
