% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phmmer_2abl_A}
\alias{phmmer_2abl_A}
\title{phmmer results after searching for 2abl_A}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://hmmer.org/}
}
\usage{
phmmer_2abl_A
}
\description{
A dataset containing
}
\keyword{datasets}
