% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_specific_conservation_data.R
\name{position_specific_conservation_data}
\alias{position_specific_conservation_data}
\title{Position-specific conservation for each position}
\usage{
position_specific_conservation_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Data Frame
}
\description{
Position-specific conservation for each position
}
