% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_specific_conservation_data.R
\name{position_specific_conservation_plot}
\alias{position_specific_conservation_plot}
\title{Plot the position-specific conservation values for each position.}
\usage{
position_specific_conservation_plot(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
Plot the position-specific conservation values for each position.
}
