% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_request_to_phmer.R
\name{post_request_to_phmer}
\alias{post_request_to_phmer}
\title{post_request_to_phmer}
\usage{
post_request_to_phmer(body_list)
}
\arguments{
\item{body_list}{A list with body parameters for the POST request.}
}
\value{
A URL with the results.
}
\description{
post_request_to_phmer
}
\examples{
\dontrun{
post_request_to_phmer(body_list =
    list(seqdb = "pdb",seq = '>Seq\nKLRVLGYH'))
}
}
