% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pySCA.R
\docType{data}
\name{pySCA_HBG2_HUMAN}
\alias{pySCA_HBG2_HUMAN}
\title{pySCA_HBG2_HUMAN}
\format{
An object of class \code{pySCA_pickle} of length 3.
}
\source{
\url{https://ranganathanlab.gitlab.io/}
}
\usage{
pySCA_HBG2_HUMAN
}
\description{
It is a list equivalent to the \code{db} object, and it contains three lists, called
sequences, sca and sector.
}
\keyword{datasets}
