% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_AA_search_using_phmmer.R
\name{quick_AA_search_using_phmmer}
\alias{quick_AA_search_using_phmmer}
\title{quick_AA_search_using_phmmer}
\usage{
quick_AA_search_using_phmmer(seqaa, seqdb)
}
\arguments{
\item{seqaa}{A character vector with AA sequences}

\item{seqdb}{A character vector with target databases.}
}
\value{
A df containing url's for downloading different files.
}
\description{
quick_AA_search_using_phmmer
}
\examples{
\dontrun{
xml.document <- get_phmer_results_from_seq("DPNLFVALYDFVASGDNTLSIT", "pdb")
}
}
