% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pySCA_pickle.R
\name{read_pySCA_pickle}
\alias{read_pySCA_pickle}
\title{Reads the pickle file produced by pySCA and returns a list of class pySCA_pickle.}
\usage{
read_pySCA_pickle(path)
}
\arguments{
\item{path}{Pat A character with the path file.}
}
\value{
pySCA_pickle
}
\description{
Reads the pickle file produced by pySCA and returns a list of class pySCA_pickle.
}
\examples{
\dontrun{
pysca <- import_pySCA_module_using_reticulate( virtualenv = "r-pySCA", force_installing = FALSE)
path <- "inst/extdata/HBG2_HUMAN.ClustalW.db"
pySCA_HBG2_HUMAN <- read_pySCA_pickle(path)
}
}
