% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCA_correlation_matrix_data.R
\name{SCA_correlation_matrix_data}
\alias{SCA_correlation_matrix_data}
\title{SCA correlation matrix data}
\usage{
SCA_correlation_matrix_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Data Frame
}
\description{
SCA correlation matrix data
}
