% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SCA_positional_coevolution_matrix_components_data.R
\name{SCA_positional_coevolution_matrix_components_data}
\alias{SCA_positional_coevolution_matrix_components_data}
\title{It returns the independent components and the eigenmodes (PCA) of Csca.}
\usage{
SCA_positional_coevolution_matrix_components_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Data Frame
}
\description{
It returns the independent components and the eigenmodes (PCA) of Csca.
}
