% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/SCA_positional_coevolution_matrix_eigenvalues_data.R
\name{SCA_positional_coevolution_matrix_eigenvalues_data}
\alias{SCA_positional_coevolution_matrix_eigenvalues_data}
\title{Eigenspectrum of the SCA positional coevolution matrix and 10 trials of matrix randomization for comparison}
\usage{
SCA_positional_coevolution_matrix_eigenvalues_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Data Frame
}
\description{
Eigenspectrum of the SCA positional coevolution matrix and 10 trials of matrix randomization for comparison
}
