% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_correlation_matrix_data.R
\name{sequence_correlation_matrix_plot}
\alias{sequence_correlation_matrix_plot}
\title{Plot a facet grid projection}
\usage{
sequence_correlation_matrix_plot(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
Plot a facet grid projection
}
