% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_correlation_matrix_projections_data.R
\name{sequence_correlation_matrix_projections_data}
\alias{sequence_correlation_matrix_projections_data}
\title{It computes the data from the PCA and IC and creates a tidy DataFrame
in order to visualize the projections of the sequence correlation matrix.}
\usage{
sequence_correlation_matrix_projections_data(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\value{
Data Frame
}
\description{
It computes the data from the PCA and IC and creates a tidy DataFrame
in order to visualize the projections of the sequence correlation matrix.
}
