% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_correlation_matrix_projections_data.R
\name{sequence_correlation_matrix_projections_plot}
\alias{sequence_correlation_matrix_projections_plot}
\title{Plot a facet grid projection}
\usage{
sequence_correlation_matrix_projections_plot(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
Plot a facet grid projection
}
