% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_sequence_identities_data.R
\name{pairwise_sequence_identities_hist}
\alias{pairwise_sequence_identities_hist}
\title{It plots pairwise sequence identities}
\usage{
pairwise_sequence_identities_hist(db)
}
\arguments{
\item{db}{Pickle pySCA}
}
\description{
It plots pairwise sequence identities
}
