% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_AA_search_using_phmmer.R
\name{quick_AA_search_using_phmmer}
\alias{quick_AA_search_using_phmmer}
\title{quick_AA_search_using_phmmer}
\usage{
quick_AA_search_using_phmmer(seq, seqdb)
}
\arguments{
\item{seq}{A string of an AA sequence}

\item{seqdb}{A string of target database.}
}
\value{
A xml_document containig the results from phmmer.
}
\description{
quick_AA_search_using_phmmer
}
\examples{
\dontrun{
xml.document <- get_phmer_results_from_seq("DPNLFVALYDFVASGDNTLSIT", "pdb")
}
}
